Phase 3b: browse every tag exactly once, and stop the cursor wrapping
Closes F7 from the adversarial review.

Browse skipped tags and could loop. Three separate causes.

The adapter truncated a device page with .take(max) but returned the device's
own cursor, which follows the last record of the FULL page - so every record it
discarded was skipped forever. A truncated page now resumes from the last
record actually returned; only an untruncated page passes the device's cursor
through.

list_tags carried a u16 cursor while symbol instance ids are u32, and computed
the next cursor as (id & 0xFFFF) + 1. On a controller with instance ids above
65535 the mask wrapped the cursor back toward 1 and the walk ran forever. The
cursor is u32 end to end now, the mask is gone, and paging past 0xFFFF emits
the 32-bit logical instance segment (0x26), pinned by a hand-assembled golden
vector since no bench peer serves that many symbols.

An invalid cursor was swallowed and the walk silently restarted at instance 1.
It is a typed refusal now, before any I/O reaches the device.

Guards where a hostile or broken device could hang us: a page that does not
advance the cursor, a non-numeric device cursor, and a page cap on the
hierarchical walk - which previously hung forever.

Review found two more. A page whose records are not ascending still skipped
silently: the resume is last.instance_id + 1, so a page of [10, 2, 3] returned
TAG_10 and resumed at 11, dropping 2 and 3 - the same defect this PR exists to
eliminate. list_tags now rejects a non-ascending or duplicated page outright.
And a mid-walk refusal reported BROWSE_UNSUPPORTED, claiming the device has no
tag-list service, when page 1 had just come from that very service; a refused
resume is now BROWSE_FAILED and says so.

Also here: an uncursored browse started at symbol instance 1, silently skipping
anything at instance 0 - the same class again, and the reason sb/browse could
not reach EthernetIPSharp at all, since that peer serves its table from 0. It
starts at 0 now. This overrides the paging spec on evidence: every enip live
suite already walked from 0, and the adapter was the sole outlier.

Carried forward: DeviceRegistry::handle/sole_handle replace a full
routing-snapshot clone per request, and the reload rollback loop keeps a
comment recording why it is unreachable rather than looking like dead code.

Bench, against the live EthernetIPSharp peer. At the enip layer the walk
enumerates all seven tags exactly once in ascending order, and a cursor above
0xFFFF emits 0x26 which the peer parses - answering "no such instance", not a
path error. At the adapter layer page 1 is captured: no cursor, max 2 ->
["LINE_SPEED", "FILL_TEMP"], next "4" - the instance-0 start and a truncation
cursor taken from the last record returned rather than the device's. Page 2 is
not obtainable: this peer refuses every non-zero start instance, so it cannot
resume at any page size. The multi-page adapter walk stays on the real-Logix
lab row and is claimed nowhere as proven.

Gates: 645 tests, clippy clean with and without tls.

+++ crates/enip/src/cip/services.rs
@@ -30,7 +30,7 @@ impl EipClient {
     ) -> Result<Bytes> {
         let path = EPath::new()
             .class(class)
-            .instance(instance)
+            .instance(u32::from(instance))
             .attribute(attribute);
         let mr = MessageRequest::new(SERVICE_GET_ATTRIBUTE_SINGLE, path, Bytes::new());
         let reply = self.send_cip(mr, "get_attribute_single").await?;
@@ -52,7 +52,7 @@ impl EipClient {
     ) -> Result<()> {
         let path = EPath::new()
             .class(class)
-            .instance(instance)
+            .instance(u32::from(instance))
             .attribute(attribute);
         let mr = MessageRequest::new(SERVICE_SET_ATTRIBUTE_SINGLE, path, data);
         let reply = self.send_cip(mr, "set_attribute_single").await?;
@@ -67,7 +67,7 @@ impl EipClient {
     /// `Get_Attribute_All` (§7.5) against `(class, instance)` — returns the concatenated attribute
     /// block for the caller to slice per the object definition.
     pub async fn get_attribute_all(&self, class: u16, instance: u16) -> Result<Bytes> {
-        let path = EPath::new().class(class).instance(instance);
+        let path = EPath::new().class(class).instance(u32::from(instance));
         let mr = MessageRequest::new(SERVICE_GET_ATTRIBUTE_ALL, path, Bytes::new());
         let reply = self.send_cip(mr, "get_attribute_all").await?;
         reply.expect_service(SERVICE_GET_ATTRIBUTE_ALL)?;
